refactor: extract get_type_alias_info cluster to QueryEngine (Phase 1)
Move get_type_alias_info and 4 helpers from CppAnalyzer to QueryEngine:
- _get_alias_details_from_db
- _get_info_for_known_alias
- _find_type_matches
- _check_type_ambiguity
- get_type_alias_info

CppAnalyzer keeps 1-line delegation for backward compatibility.
All state access in QueryEngine goes through self.analyzer back-reference.

All 1435 tests pass.
