+++ mcp_server/cpp_analyzer.py
@@ -1525,177 +1525,9 @@ class CppAnalyzer:
         """Get signature details for functions (delegates to query_engine)."""
         return self.query_engine.get_function_signature(function_name, class_name)
 
-    def _get_alias_details_from_db(self, alias_names: List[str]) -> List[Dict[str, Any]]:
-        """Query type_aliases table for detailed information about a set of aliases."""
-        unique_aliases = {}
-        try:
-            self.cache_manager.backend._ensure_connected()
-            conn = self.cache_manager.backend.conn
-            assert conn is not None
-
-            for alias_name in alias_names:
-                cursor = conn.execute(
-                    """
-                    SELECT alias_name, qualified_name, canonical_type, file, line, namespace,
-                           is_template_alias, template_params
-                    FROM type_aliases
-                    WHERE alias_name = ? OR qualified_name = ?
-                    """,
-                    (alias_name, alias_name),
-                )
-                row = cursor.fetchone()
-                if row:
-                    qualified_alias = row["qualified_name"]
-                    if qualified_alias not in unique_aliases:
-                        alias_dict = {
-                            "name": row["alias_name"],
-                            "qualified_name": qualified_alias,
-                            "file": row["file"],
-                            "line": row["line"],
-                        }
-                        if row["is_template_alias"]:
-                            alias_dict["is_template_alias"] = True
-                            if row["template_params"]:
-                                alias_dict["template_params"] = json.loads(row["template_params"])
-                        unique_aliases[qualified_alias] = alias_dict
-        except Exception as e:
-            diagnostics.debug(f"Failed to get alias details: {e}")
-        return list(unique_aliases.values())
-
-    def _get_info_for_known_alias(self, type_name: str) -> Optional[Dict[str, Any]]:
-        """Attempt to get type alias info from the database if type_name is a known alias."""
-        try:
-            self.cache_manager.backend._ensure_connected()
-            conn = self.cache_manager.backend.conn
-            assert conn is not None
-            cursor = conn.execute(
-                """
-                SELECT alias_name, qualified_name, canonical_type, file, line, namespace,
-                       is_template_alias, template_params
-                FROM type_aliases
-                WHERE alias_name = ? OR qualified_name = ?
-                """,
-                (type_name, type_name),
-            )
-            row = cursor.fetchone()
-            if row:
-                alias_names = self.cache_manager.get_aliases_for_canonical(row["canonical_type"])
-                aliases = self._get_alias_details_from_db(alias_names)
-
-                return {
-                    "canonical_type": row["canonical_type"],
-                    "qualified_name": row["qualified_name"],
-                    "namespace": row["namespace"],
-                    "file": row["file"],
-                    "line": row["line"],
-                    "input_was_alias": True,
-                    "is_ambiguous": False,
-                    "aliases": aliases,
-                }
-        except Exception as e:
-            diagnostics.warning(f"Error querying type_aliases for '{type_name}': {e}")
-        return None
-
-    def _find_type_matches(self, type_name: str) -> List[SymbolInfo]:
-        """Search class index for matching types and return list of matches."""
-        matches = []
-        with self.index_lock:
-            for name, infos in self.class_index.items():
-                for info in infos:
-                    # Use qualified pattern matching (same as search_classes)
-                    qualified_name = info.qualified_name if info.qualified_name else info.name
-                    if SearchEngine.matches_qualified_pattern(qualified_name, type_name):
-                        matches.append(info)
-        return matches
-
-    def _check_type_ambiguity(
-        self, type_name: str, matches: List[SymbolInfo]
-    ) -> Optional[Dict[str, Any]]:
-        """Check for ambiguity among matches and return error dict if ambiguous."""
-        if len(matches) > 1:
-            # Check if all matches have the same qualified_name (duplicates from forward decls)
-            unique_qualified_names = set(
-                m.qualified_name if m.qualified_name else m.name for m in matches
-            )
-            if len(unique_qualified_names) > 1:
-                # Ambiguous - multiple different types match
-                return {
-                    "error": f"Ambiguous type name '{type_name}'",
-                    "is_ambiguous": True,
-                    "matches": [
-                        {
-                            "canonical_type": m.name,
-                            "qualified_name": m.qualified_name if m.qualified_name else m.name,
-                            "namespace": m.namespace,
-                            "file": m.file,
-                            "line": m.line,
-                        }
-                        for m in matches
-                    ],
-                    "suggestion": "Use qualified name (e.g., 'ui::Widget')",
-                }
-        return None
-
     def get_type_alias_info(self, type_name: str) -> Dict[str, Any]:
-        """Get comprehensive type alias information."""
-        # Step 1: Check if input is a known alias in the type_aliases table
-        input_canonical = self.cache_manager.get_canonical_for_alias(type_name)
-        input_was_alias = False
-
-        if input_canonical:
-            input_was_alias = True
-            info = self._get_info_for_known_alias(type_name)
-            if info:
-                return info
-
-        # Step 2: Input is not a known alias - search class_index for matching classes
-        matches = self._find_type_matches(type_name)
-
-        # Step 3: Check for ambiguity
-        ambiguity_error = self._check_type_ambiguity(type_name, matches)
-        if ambiguity_error:
-            return ambiguity_error
-
-        # Step 4: Handle not found case
-        if len(matches) == 0:
-            return {
-                "error": f"Type '{type_name}' not found",
-                "canonical_type": None,
-                "aliases": [],
-            }
-
-        # Step 5: Use the first match (prefer definitions over declarations)
-        canonical_info = matches[0]
-        for m in matches:
-            if m.is_definition:
-                canonical_info = m
-                break
-
-        canonical_type = (
-            canonical_info.qualified_name if canonical_info.qualified_name else canonical_info.name
-        )
-
-        # Step 6: Get all aliases for the canonical type
-        alias_names = self.cache_manager.get_aliases_for_canonical(canonical_type)
-
-        # Step 7: Build detailed alias information
-        aliases = self._get_alias_details_from_db(alias_names) if alias_names else []
-
-        # Step 8: Return success result
-        return {
-            "canonical_type": canonical_type,
-            "qualified_name": (
-                canonical_info.qualified_name
-                if canonical_info.qualified_name
-                else canonical_info.name
-            ),
-            "namespace": canonical_info.namespace,
-            "file": canonical_info.file,
-            "line": canonical_info.line,
-            "input_was_alias": input_was_alias,
-            "is_ambiguous": False,
-            "aliases": aliases,
-        }
+        """Get comprehensive type alias information (delegates to query_engine)."""
+        return self.query_engine.get_type_alias_info(type_name)
 
     def search_symbols(
         self,
